refactor(mitm-addon): split firewall matching helpers
